pyproject.toml
[tool.poetry]
name = "pybox"
version = "0.1.0"
description = "Just a sandbox for Python stuff"
authors = ["margostino"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.1"
beautifulsoup4 = "^4.12.3"
itemadapter = "^0.8.0"
jinja2 = "^3.1.3"
pyyaml = "^6.0.1"
requests = "^2.31.0"
scrapy = "^2.11.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
matplotlib = "^3.9.1"
selenium = "^4.22.0"
webdriver-manager = "^4.0.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
bs4 = "^0.0.2"
pydantic = "2.8.2"
weaviate-client = "^4.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
